MAINT: embed cell outputs in HTML again
The reason to remove cell output for matplotlib plots was that they
don't render in the PDF build. Instead, the cell output is removed and
the figure is exported and rendered through Markdown. The disadvantage
of this setup is that it conflicts with MyST-NB caching: if you switch
commits and regenerate documentation, the rendered figures are not
fetched from cache, but are just inserted from the latest run.

This commit takes for granted that texts like
```
<Figure 600x600 ... >
```
appear in the PDF build and reverts the old caching behavior.